% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_estimation.R
\name{.single.TNBsurvival}
\alias{.single.TNBsurvival}
\title{Estimate Survival for Specific Times}
\usage{
.single.TNBsurvival(data, times)
}
\arguments{
\item{data}{A data frame with survival estimates (output of \code{.process_single_run}).}

\item{times}{Numeric vector of times for which survival estimates are required.}
}
\value{
A numeric vector of survival probabilities corresponding to the input times.
}
\description{
This function estimates survival probabilities for a given set of times using interval-censored data.
}
\keyword{internal}
